# oas/simaas_oas2.json
{
    "swagger": "2.0",
    "info": {
        "title": "SIMaaS Using Queue, Workers and fmpy",
        "version": "0.3.0"
    },
    "paths": {
        "/experiments": {
            "post": {
                "tags": ["Experiments"],
                "summary": "Trigger the simulation of a model instance by defining an experiment",
                "consumes": [ "application/json" ],
                "parameters": [
                    {
                        "$ref": "#/parameters/ExperimentSetup"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "An acknowledgement that the simulation was successfully triggered",
                        "headers": {
                            "Location": {
                                "description": "Referral to a resource that indicates the status of the long-running simulation; this resource needs to be polled ([RFC7231, section 7.1.2](https://tools.ietf.org/html/rfc7231#section-7.1.2))",
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/experiments/{uuid}/status": {
            "get": {
                "tags": ["Experiments"],
                "summary": "A resource indicating the status of an experiment",
                "consumes": [ "application/json" ],
                "produces": [ "application/json" ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/ExperimentStatus"
                    }
                }
            },
            "parameters": [
                { "$ref": "#/parameters/ExperimentID" }
            ]
        },
        "/experiments/{uuid}/result": {
            "get": {
                "tags": ["Experiments"],
                "summary": "The results of performing the experiment/simulation",
                "consumes": [ "application/json" ],
                "produces": [ "application/json" ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/SimulationResult"
                    }
                }
            },
            "parameters": [
                { "$ref": "#/parameters/ExperimentID" }
            ]
        }
    },
    "definitions": {
        "SimulationParameters": {
            "type": "object",
            "required": [
                "startTime",
                "stopTime"
            ],
            "properties": {
                "startTime": {
                    "description": "The beginning of the desired time frame for the simulation in [milliseconds since the UNIX epoch](https://en.wikipedia.org/wiki/Unix_time)",
                    "type": "integer",
                    "format": "int64",
                    "example": "1526637990000"
                },
                "stopTime": {
                    "description": "The end of the desired time frame for the simulation in [milliseconds since the UNIX epoch](https://en.wikipedia.org/wiki/Unix_time)",
                    "type": "integer",
                    "format": "int64",
                    "example": "1526551590000"
                },
                "outputInterval": {
                    "description": "The desired output interval in seconds (does _not_ affect step size of the integration algorithm)",
                    "type": "number",
                    "example": "900"
                }
            }
        },
        "InputTimeseries": {
            "type":"array",
            "items": [
                {
                    "allOf": [
                        {
                            "$ref":"https://raw.githubusercontent.com/UdSAES/designetz_schemata/master/schemata/timeseries/schema_v1.2.0-oas2.json#/Timeseries"
                        },
                        {
                            "required": [
                                "label"
                            ]
                        },
                        {
                            "properties": {
                                "label": {
                                    "type": "string",
                                    "enum": [ "SaaSu" ]
                                }
                            }
                        }
                    ]
                },
                {
                    "allOf":[
                        {
                            "$ref":"https://raw.githubusercontent.com/UdSAES/designetz_schemata/master/schemata/timeseries/schema_v1.2.0-oas2.json#/Timeseries"
                        },
                        {
                            "required": [ "label" ]
                        },
                        {
                            "properties": {
                                "label": {
                                    "type":"string",
                                    "enum": [ "SaaSu1" ]
                                }
                            }
                        }
                    ]
                }
            ],
            "minItems":2,
            "maxItems":2
        },
        "ModelInstanceID": {
            "description": "The UUID of the model instance to be simulated",
            "type": "string",
            "enum": [ "c02f1f12-966d-4eab-9f21-dcf265ceac71" ]
        }
    },
    "parameters": {
        "ExperimentSetup": {
            "name": "experimentSetup",
            "in": "body",
            "description": "The boundary conditions and simulation parameters that define an experiment",
            "required": true,
            "schema": {
                "type": "object",
                "required": [
                    "modelInstanceID",
                    "simulationParameters"
                ],
                "properties": {
                    "modelInstanceID": {
                        "$ref": "#/definitions/ModelInstanceID"
                    },
                    "simulationParameters": {
                        "$ref": "#/definitions/SimulationParameters"
                    },
                    "inputTimeseries": {
                        "$ref": "#/definitions/InputTimeseries"
                    }
                },
                "example": {"modelInstanceID":"c02f1f12-966d-4eab-9f21-dcf265ceac71","simulationParameters":{"startTime":1526558400000,"stopTime":1526569200000,"outputInterval":900},"inputTimeseries":[{"label":"SaaSu","unit":"W/m2","timeseries":[{"timestamp":1526558400000,"value":555.125},{"timestamp":1526559300000,"value":551.1875},{"timestamp":1526560200000,"value":568.28125},{"timestamp":1526561100000,"value":628.59375},{"timestamp":1526562000000,"value":631.875},{"timestamp":1526562900000,"value":607.7968750000724},{"timestamp":1526563800000,"value":593.4218750190755},{"timestamp":1526564700000,"value":559.8906250641351},{"timestamp":1526565600000,"value":515.2031250833663},{"timestamp":1526566500000,"value":458.42187504998253},{"timestamp":1526567400000,"value":401.109375043708},{"timestamp":1526568300000,"value":360.6250000271315},{"timestamp":1526569200000,"value":338.6562500432486}]},{"label":"SaaSu1","unit":"W/m2","timeseries":[{"timestamp":1526558400000,"value":238.53135991096497},{"timestamp":1526559300000,"value":238.43617391586304},{"timestamp":1526560200000,"value":220.99572658538818},{"timestamp":1526561100000,"value":171.54822373390198},{"timestamp":1526562000000,"value":158.77959752082825},{"timestamp":1526562900000,"value":161.28478050231934},{"timestamp":1526563800000,"value":157.22860145568848},{"timestamp":1526564700000,"value":164.33479499816895},{"timestamp":1526565600000,"value":176.16250228881836},{"timestamp":1526566500000,"value":193.0199851989746},{"timestamp":1526567400000,"value":207.74025344848633},{"timestamp":1526568300000,"value":210.68580627441406},{"timestamp":1526569200000,"value":201.1289520263672}]}]}
            }
        },
        "ExperimentID": {
            "name": "uuid",
            "in": "path",
            "description": "A unique identifier for a specific experiment",
            "required": true,
            "type": "string",
            "x-example": "d9a23e72-b01a-4be7-a598-c3f31071da2c"
        }
    },
    "responses": {
        "ExperimentStatus": {
            "description": "The current status of the experiment",
            "schema": {
                "type": "object",
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [ "NEW", "IN_PROGRESS", "DONE" ],
                        "description": "The status of the background job, for example 'IN_PROGRESS'"
                    },
                    "linkToResult": {
                        "type": "string",
                        "description": "The link to the initially requested resource (the simulation result) that is now available. This property only exists if the status has reached 'DONE' and the result is indeed available."
                    }
                }
            },
            "examples": {
                "application/json": { "status": "DONE", "linkToResult": "https://host.tld/experiments/uuid/result"}
            }
        },
        "SimulationResult": {
            "description": "The simulation result",
            "schema": {
                "type": "object",
                "required": [
                    "description",
                    "result"
                ],
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "result": {
                        "type": "array",
                        "items": {
                            "$ref": "https://raw.githubusercontent.com/UdSAES/designetz_schemata/master/schemata/timeseries/schema_v1.2.0-oas2.json#/Timeseries"
                        }
                    }
                }
            },
            "examples": {
                "application/json": {"description": "An exemplary simulation result", "data": [ { "label": "albedo_ground", "unit": "1", "timeseries": [ { "timestamp":
                1526634000000, "value": 0.15174804687500001 }, { "timestamp": 1526637600000,
                "value": 0.15174804687500001 } ] }, { "label": "temperature_2m_ag", "unit": "K",
                "timeseries": [ { "timestamp": 1526634000000, "value": 286.0401916503906 }, {
                "timestamp": 1526637600000, "value": 287.0859680175781 } ] } ]}
            }
        }
    },
    "tags": [
        {
            "name": "Model Instances",
            "description": "A model instance is the executable version of a model with all parameters set, thus describing the behaviour of a specific system for all possible boundary conditions. The parameters of a model instance cannot be changed in order to ensure that identical service interactions always lead to the same result. Consequently, it is possible that different model instances describe the same specific system, which can be a result of different parameter sets or the use of different models."
        },
        {
            "name": "Experiments",
            "description": "An experiment denotes the execution of a model instance for given boundary conditions and simulation settings; in other words, a specific simulation."
        }
    ]
}